+++ packages/website/src/_services/api/taxonomy-service/taxonomy-service-mock.ts
@@ -1,8 +1,8 @@
 import { TaxonomyClass } from '../types'
 
 export const TAXONOMY_CLASS_ALL: TaxonomyClass = { name: 'All', symbol: 'Σ' }
+
 export const TAXONOMY_CLASSES: TaxonomyClass[] = [
-  TAXONOMY_CLASS_ALL,
   { name: 'Amphibians', symbol: '🐸' },
   { name: 'Birds', symbol: '🐦' },
   { name: 'Mammals', symbol: '🐗' }