+++ packages/website/src/_services/maps/map-tool-menu/map-tool-menu.ts
@@ -1,8 +1,7 @@
 import { Vue } from 'vue-class-component'
-import { Emit, Prop, Watch } from 'vue-property-decorator'
+import { Emit, Prop } from 'vue-property-decorator'
 
 import { MAPBOX_STYLE_RFCX_WITH_PLACE_LABELS, MAPBOX_STYLE_SATELLITE_STREETS, MapboxStyle } from '~/maps'
-import { TAXONOMY_CLASSES } from '../../api/taxonomy-service'
 
 interface MapOptions {
   style: MapboxStyle
@@ -10,13 +9,7 @@ interface MapOptions {
 }
 
 export default class MapToolMenuComponent extends Vue {
-  @Prop({ default: false }) displayTaxonomies!: boolean
   @Prop() mapStyle!: MapboxStyle
-
-  @Emit() @Watch('taxon')
-  emitTaxonomyValue (): string {
-    return this.taxon
-  }
 
   @Emit()
   emitShowLabelsToggle (): boolean {
@@ -29,8 +22,6 @@ export default class MapToolMenuComponent extends Vue {
     return style
   }
 
-  taxons = TAXONOMY_CLASSES
-  taxon = this.taxons[0].name
   isShowLabels = true
   mapStyleOptions: MapOptions[] = [
     { style: MAPBOX_STYLE_SATELLITE_STREETS, name: 'Satellite' },