+++ packages/website/src/species-richness/components/species-richness-by-location/species-richness-by-location.ts
@@ -2,7 +2,7 @@ import { Options, Vue } from 'vue-class-component'
 import { Prop } from 'vue-property-decorator'
 
 import { generateHtmlPopup } from '@/species-richness/components/species-richness-by-location/functions'
-import { TAXONOMY_CLASSES } from '~/api/taxonomy-service'
+import { TAXONOMY_CLASS_ALL } from '~/api/taxonomy-service'
 import { getExportFilterName } from '~/dataset-filters/functions'
 import { DEFAULT_LATITUDE, DEFAULT_LONGITUDE, MAPBOX_STYLE_SATELLITE_STREETS, MapboxStyle } from '~/maps'
 import { MapBubbleComponent, MapConfig, MapDataSet } from '~/maps/map-bubble'
@@ -19,9 +19,6 @@ const DEFAULT_PREFIX = 'Species-By-Site'
 export default class SpeciesRichnessByLocation extends Vue {
   @Prop({ default: [] }) public datasets!: MapDataSet[]
 
-  taxon = TAXONOMY_CLASSES[0].name
-  taxons = TAXONOMY_CLASSES
-
   isShowLabels = true
   mapStyle: MapboxStyle = MAPBOX_STYLE_SATELLITE_STREETS // TODO: Encapsulate this under BubbleMapGroup
   getPopupHtml = generateHtmlPopup
@@ -30,6 +27,10 @@ export default class SpeciesRichnessByLocation extends Vue {
     sourceMapId: '',
     center: [DEFAULT_LONGITUDE, DEFAULT_LATITUDE],
     zoom: 9
+  }
+
+  get mapDataKey (): string {
+    return TAXONOMY_CLASS_ALL.name
   }
 
   get hasData (): boolean {
@@ -43,7 +44,6 @@ export default class SpeciesRichnessByLocation extends Vue {
     }
   }
 
-  propagateTaxonomyValue (taxon: string): void { this.taxon = taxon }
   propagateMapMove (config: MapConfig): void { this.config = config }
   propagateMapStyle (style: MapboxStyle): void { this.mapStyle = style }
   propagateToggleLabels (isShowLabels: boolean): void { this.isShowLabels = isShowLabels }