+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Allowed categories:
 
 - **activity-overview:** Group detection, detection frequency, and occupancy map [#255][#257]
 - **activity-overview:** Group detection, detection frequency, and occupancy map [#259][#260]
+- **selection:** Add taxon in dataset comparison filter [#219]
+- **selection:** Remove taxon toggle from map [#267]
 
 # 0.0.8 (2021-11-22)